+++ client/src/components/NewActivity/use-new-activity.ts
@@ -1,7 +1,7 @@
+import useRouteProps from "@/lib/useRouteProps";
 import { hasUserId } from "@/types/server/user-id.guards";
 import { useTagSelection } from "@lib/state/selected-tags-state";
 import useAuthentication from "@lib/use-authentication";
-import useRouteProps from "@lib/use-route-props";
 import type { DateTimeField } from "@type/form.types";
 import type { NewActivity } from "@type/server/activity.types";
 import { useEffect, useMemo, useState } from "react";
