+++ client/src/components/NewNote/use-new-note.ts
@@ -1,9 +1,9 @@
+import useRouteProps from "@/lib/useRouteProps";
 import { queryClient } from "@lib/query-client";
 import { useNewNoteMutation } from "@lib/query/use-new-note-mutation";
 import useTagsQuery from "@lib/query/use-tags-query";
 import { useTagSelection } from "@lib/state/selected-tags-state";
 import useAuthentication from "@lib/use-authentication";
-import useRouteProps from "@lib/use-route-props";
 import type { NewNote } from "@type/server/note.types";
 import { useEffect, useState } from "react";
